requirements.txt
pandas==2.2.2
streamlit==1.37.1
langchain-community==0.2.12
google-generativeai==0.7.2
python-dotenv==1.0.1
transformers==4.44.1
faiss-cpu==1.8.0.post1
torch==2.4.0
